+++ src/token.rs
@@ -3,12 +3,12 @@ use std::fmt::{Display, Formatter};
 #[derive(Debug, Clone, PartialEq)]
 pub struct Token<'src> {
     pub token_type: TokenType,
-    pub line: usize,
+    pub line: u32,
     pub(crate) lexeme: &'src str,
 }
 
 impl<'src> Token<'src> {
-    pub fn new(token_type: TokenType, line: usize, lexeme: &'src str) -> Self {
+    pub fn new(token_type: TokenType, line: u32, lexeme: &'src str) -> Self {
         Self {
             token_type,
             line,
